I don't think I want to support TYPES
Having an object in perl6 instead of a scalar value in perl5 will make
types easier on the user side.

What I *do* envision is the possibility to allow to subclass CSV::Field
and tell Text::CSV to use the specified class instead of CSV::Field, so
the end-user has *ALL* the flexibility he/she wants/needs.

Not a top priority though