Move test-t.pl -> test-tuxic.pl
Testing Text::CSV as a "normal" user, shouldn't need to incur the grammar
rewriting overhead of Slang::Tuxic.  So to get a better idea of the
performance of Text::CSV for "normal" users, it feels like a good idea
to have test-t.pl be in "generic" Perl 6.  To allow testing *with*
Slang::Tuxic, we now have the test-tuxic.pl file (which contains the
original version of test-t.pl).